optimize expensive re-renders
